config and reaction roles (WIP)
Added a config.cfg file to modify easily the ID of channels and other things, that are changeable. Added a log, for a channel of your wish, to which the bot sends a message, when he is online. This can be disabled by setting the log channel 0 in the .cfg. Still working on a reaction role function, to verfiy your members after reading the rules.